refactor: split CtRegionMap state and viewport input parts (#3878)
Extract lifecycle/bus/tap handlers and keyboard/scroll/hover/pinch input
chrome into library parts; document paths in repo-and-packages.

Refs #3878
